+++ workspace/solutions/kotlin/lib/src/test/kotlin/s0/s0/s5/s6/s_0056_test.kt
@@ -1,0 +1,16 @@
+package s0.s0.s5.s6
+
+import utils.T
+import kotlin.test.Test
+import s0.s0.s5.s6.s_0056_00.Solution as s_0056_00
+
+class s_0056_test {
+    @Test
+    fun test() {
+        val t = T(
+            this::class,
+            s_0056_00::class,
+        )
+        t.run()
+    }
+}