+++ workspace/solutions/rust/src/s0/s0/mod.rs
@@ -1,3 +1,4 @@
 pub mod s0;
 pub mod s3;
 pub mod s4;
+pub mod s5;