+++ workspace/solutions/rust/src/s0/s0/s5/s6/mod.rs
@@ -1,0 +1,4 @@
+pub mod s_0056_00;
+pub mod s_0056_test;
+
+pub type TSolution = fn(Vec<Vec<i32>>) -> Vec<Vec<i32>>;